Typing

+++ homeassistant/components/ambient_station/__init__.py
@@ -43,7 +43,7 @@ CONFIG_SCHEMA = cv.deprecated(DOMAIN)
 
 
 @callback
-def async_wm2_to_lx(value: float) -> float:
+def async_wm2_to_lx(value: float) -> int:
     """Calculate illuminance (in lux)."""
     return round(float(value) / 0.0079)
